Do not allow SLAAC on non-mac devices

+++ modules/pico_slaacv4.h
@@ -11,7 +11,7 @@
 #define PICO_SLAACV4_SUCCESS  0
 #define PICO_SLAACV4_ERROR    1
 
-uint8_t pico_slaacv4_claimip(struct pico_device *dev, void (*cb)(struct pico_ip4 *ip,  uint8_t code));
+int     pico_slaacv4_claimip(struct pico_device *dev, void (*cb)(struct pico_ip4 *ip,  uint8_t code));
 void    pico_slaacv4_unregisterip(void);
 
 #endif /* _INCLUDE_PICO_SUPPORT_SLAACV4 */
